Fix image build for Virtual Box support
This change fixes the image build, which creates a bootable image for
the game. Since the sprites were too large, they and the game could not
fit on a floppy disk anymore, so it now builds a 16 bit bootable FAT image,
which should be inserted as CD-ROM.

This change allows the game to run on Virtual Box, but it seems that the
game does not display its sprites correctly on Virtual Box, which makes
the game unplayable on Virtual Box for now.

This will hopefully addressed in the future.